move tests to meta/reflect pkg
